# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Scooter API Document",
    "description": "test",
    "version": "1.0.0"
  },
  "host": "http://localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "성공"
          }
        }
      }
    },
    "/user/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "성공"
          }
        }
      }
    },
    "/user/verify": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "401": {
            "description": "권한 없음"
          },
          "419": {
            "description": ""
          }
        }
      }
    },
    "/user/checkid": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "409": {
            "description": "충돌"
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                },
                "userPw": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "401": {
            "description": "권한 없음"
          },
          "403": {
            "description": "금지됨"
          }
        }
      }
    },
    "/user/signup": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "성공"
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
    "/user/users": {
      "get": {
        "tags": [
          "get"
        ],
        "description": "유저 전체 조회",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "user": {
                  "type": "string",
                  "example": "testuser"
                },
                "password": {
                  "type": "string",
                  "example": "test1234"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "201": {
            "description": "유저 전체 조회 성공",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "유저 리스트"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "권한 없음"
          },
          "403": {
            "description": "금지됨"
          },
          "419": {
            "description": ""
          }
        }
      }
    },
    "/user/history": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "401": {
            "description": "권한 없음"
          },
          "403": {
            "description": "금지됨"
          },
          "419": {
            "description": ""
          }
        }
      }
    },
    "/admin/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "성공"
          }
        }
      }
    },
    "/admin/update-role": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                },
                "role": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "401": {
            "description": "권한 없음"
          },
          "403": {
            "description": "금지됨"
          },
          "419": {
            "description": ""
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
    "/board/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "성공"
          }
        }
      }
    },
    "/board/generate-code": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "scooterIdx": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "401": {
            "description": "권한 없음"
          },
          "419": {
            "description": ""
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
    "/board/check-code": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "scooterIdx": {
                  "example": "any"
                },
                "code": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "401": {
            "description": "권한 없음"
          },
          "403": {
            "description": "금지됨"
          },
          "419": {
            "description": ""
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
    "/board/read-code": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "scooterIdx": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "description": "잘못된 요청"
          }
        }
      }
    },
    "/board/update-detect": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "scooterIdx": {
                  "example": "any"
                },
                "detected": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          }
        }
      }
    },
    "/board/read-verified": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "scooterIdx": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          }
        }
      }
    },
    "/board/read-alcohol": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "idx",
            "in": "query",
            "type": "string"
          },
          {
            "name": "userId",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          }
        }
      }
    },
    "/board/finish": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "scooterIdx": {
                  "example": "any"
                },
                "rentalDt": {
                  "example": "any"
                },
                "rentalStTm": {
                  "example": "any"
                },
                "rentalRtTm": {
                  "example": "any"
                },
                "payMethod": {
                  "example": "any"
                },
                "paidAmount": {
                  "example": "any"
                },
                "paidStatus": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "401": {
            "description": "권한 없음"
          },
          "419": {
            "description": ""
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
    "/board/auth-status": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "inputCode": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          }
        }
      }
    },
    "/helmet/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "성공"
          }
        }
      }
    },
    "/helmet/set-data": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "alcohol": {
                  "example": "any"
                },
                "userId": {
                  "example": "any"
                },
                "helmetIdx": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          }
        }
      }
    },
    "/pay/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "성공"
          }
        }
      }
    },
    "/pay/pay-amount": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "401": {
            "description": "권한 없음"
          },
          "419": {
            "description": ""
          }
        }
      }
    }
  }
}